More brute force for 2. 3 is finally a problem 'hard' enough for a computer that brute force takes long enough for me to want try interesting ways.
I did a for loop up to the big num initially when I guessed that if 71 was a factor, then the largest factor probably couldn't be bigger than bignum / 71, which would decrease the problem space.

The current invalid solution is an implementation of that, which doesn't result in the correct answer. I should test against the example small value given for big num to see if I am generating good factors.